+++ juniorguru/sync/interviews_tips.py
@@ -62,7 +62,8 @@ async def discord_task(client: ClubClient):
                 embeds=[embed_mentors, embed_handbook],
             )
 
-        logger.info("Deleting previous reminder")
-        message = channel.fetch_message(last_message.id)
-        with mutating_discord(message) as proxy:
-            await proxy.delete()
+        if last_message:
+            logger.info("Deleting previous reminder")
+            message = channel.fetch_message(last_message.id)
+            with mutating_discord(message) as proxy:
+                await proxy.delete()
